+++ tests/parser_spec.rs
@@ -187,9 +187,8 @@ fn should_parse_option_with_comment() {
 fn should_parse_include() {
     let includes = parse::<f64>(r#"include "./a/path/to/file.beancount""#)
         .unwrap()
-        .includes()
-        .collect::<HashSet<_>>();
-    let expected: HashSet<_> = [Path::new("./a/path/to/file.beancount")].into();
+        .includes;
+    let expected = [Path::new("./a/path/to/file.beancount")];
     assert_eq!(includes, expected);
 }
 
@@ -198,10 +197,9 @@ fn should_parse_include_with_comment() {
     let includes =
         parse::<f64>(r#"include "./a/path/to/file.beancount" ; Everything is in the other file"#)
             .unwrap()
-            .includes()
-            .collect::<HashSet<_>>();
-    let expected: HashSet<_> = [Path::new("./a/path/to/file.beancount")].into();
-    assert_eq!(includes, expected);
+            .includes;
+    let expected = &[Path::new("./a/path/to/file.beancount")];
+    assert_eq!(&includes, expected);
 }
 
 #[rstest]